Add files via upload
This version of the application swaps out the nasty for loops for array.filter and array.map functions for the working drop-down menus.  

I can start, stop, and save the swim time for the selected swimmer, but I can't yet find a way to add this time to the main swim meet object (ageGroupsFactory).